LayoutTests/fast/frames/javascript-url-anchor-target-expected.txt

Tests that javascript URL execute in the right context when set as href on an anchor with a target frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS contextName is "testFrame"
PASS successfullyParsed is true

TEST COMPLETE
